fix: actually use ccache binary returned by `find_program`
The original code assumed that if `find_program` found `ccache`,
then `ccache` must be in the `PATH`. However, `find_program`
searches in a lot of different places. This can lead to situations
where `ccache` is found, but the compile will fail because it's
not in the path.

This change also uses `CMAKE_XXX_COMPILER_LAUNCHER` instead of
`RULE_LAUNCH_COMPILE`, as the latter is for internal use only
per the CMake docs.