Support transforming stacktraces
This support is added to the stacktrace itself, rather than the
formatter, since allowing the formatter to change the stacktrace would
require *every single* `stacktrace` and `stacktrace_frame` usage to
require either an lvalue ref to non-const or an rvalue ref. This would
break the existing API, which uses lvalue refs to const for everything.

Overall, the formatter is more of a view into a stacktrace, so while it
was the first place I tried to put the transformation, it's not the best
place. Instead, transforming the stacktrace itself can be done prior to
formatting, without breaking any existing APIs.