Add move-based frame transformation
This could have been added directly in `print_frame_inner`, so we only
need to add it in one place. However, by adding it in `print_internal`,
prior to filtering, we allow transformation and filtering to cooperate.
This allows a transformer to modify a frame such that it can later be
marked for filtering, which would avoid duplicated code between those
functions.

I don't have this use case right now, but it made sense from an overall
product design perspective.