Use a versioned ABI namespace (#237)
This is something I should have done a long time ago. This is an effort
to future-proof for possible future evolution. The next release will
have an ABI break as a result of this but if it impacts any code at all
it will be a linker error rather than a silent bug.