Add cpptrace::experimental::load_symbols_for_file
For now, this function is only implemented for Windows when CPPTRACE_GET_SYMBOLS_WITH_DBGHELP is
defined. For other configurations, it does nothing.

load_symbols_for_file will load symbols from the given file with SymLoadModuleEx. It should be
called when a module was loaded at runtime with LoadLibrary. Otherwise, the symbols for frames
inside that loaded module will not be resolvable.