Change load_symbols_for_file to load_symbols_for_module
The function now takes an HMODULE handle to the DLL instead of the filename. This should make it
more explicit that this is a Windows-only function. It also reduces the (rather minor) risk of
passing a different filename into the function that doesn't actually refer to a loaded module.

It is still necessary to get the filename with GetModuleFileName, but as long as the module handle
is valid, this should rarely be an issue.