Revert to load_symbols_for_file to avoid including Windows.h
Having the HMODULE data type in the public interface requires including <Windows.h>, since there
doesn't seem to be a "proper" way of forward declaring HMODULE. However, including <Windows.h>
comes with a host of problems. For one, its sheer size. For another, it defines some popular names
like `min` as macros, which breaks `std::min` and similar.

Passing a filename instead of the module handle seems *slightly* less safe due to the risk of
passing a string that doesn't refer to a module, but it's also much less hassle.