Make `dwarf_init_path_a` error non-fatal (#251)
I ran into an exception in the case of a "corrupted" ELF section (just
unreadable by libdwarf, not corrupted). It turned out to be percolated
from `dwarf_init_path_a` returning an error. libdwarf failing to parse
an object should not inhibit falling back to parsing the ELF symtab, so
I make this particular path less fatal.

I repeated `ok = false;` for consistency and readability but I do
realize all of the appearances in the function of this line seem
redundant.

---------

Co-authored-by: Jeremy Rifkin <51220084+jeremy-rifkin@users.noreply.github.com>